Move loader to instance provider
