+++ src/app/api/thrift-instance-provider/types/api-metadata.ts
@@ -1,0 +1,5 @@
+import { ThriftAstMetadata } from '../../../thrift-services';
+
+export interface ApiMetadata {
+    [name: string]: ThriftAstMetadata[];
+}
