+++ src/app/api/thrift-instance-provider/types/index.ts
@@ -1,2 +1,3 @@
 export * from './provider-methods';
 export * from './provider-settings';
+export * from './api-metadata';
